chore: add store fetching logic to configu store
