+++ packages/common/package.json
@@ -34,7 +34,7 @@
     "@configu/ts": "workspace:*",
     "change-case": "4.1.2",
     "cosmiconfig": "^9.0.0",
-    "lodash": "^4.17.21"
-  },
-  "devDependencies": {}
+    "lodash": "^4.17.21",
+    "type-fest": "^4.8.3"
+  }
 }
